feat(backend): 예약 및 숙소 도메인 엔티티/비즈니스 로직 고도화

+++ backend/src/main/kotlin/com/mrmention/nextstay/domain/booking/entity/Booking.kt
@@ -32,7 +32,7 @@ class Booking(
     val checkOutDate: LocalDate,
 
     @Column(nullable = false)
-    val totalPrice: Int,
+    val totalPrice: Long,
 
     @Enumerated(EnumType.STRING)
     @Column(nullable = false)
